requirements.txt
certifi==2020.4.5.1
cmake==3.17.1
Distance==0.1.3
dlib==19.19.0
imutils==0.5.3
keyboard==0.13.5
numpy==1.18.3
opencv-python==4.2.0.34
playsound==1.2.2
scipy==1.4.1
wincertstore==0.2
pillow==7.1.2